src/requirements-security.txt
# Enhanced security requirements for Flask PDF API
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-cors>=6.0.1
python-dotenv>=1.1.1
PyPDF2>=3.0.1
openai>=1.95.0

# Security packages
python-magic>=0.4.27        # MIME type detection
werkzeug>=3.0.0             # Secure filename handling
cryptography>=41.0.0        # Secure cryptographic operations

# Rate limiting and validation
flask-limiter>=3.5.0        # Rate limiting
wtforms>=3.1.0              # Form validation
email-validator>=2.1.0      # Email validation

# Production server
gunicorn>=21.2.0            # WSGI server for production
gevent>=23.9.0              # Async support

# Logging and monitoring  
python-json-logger>=2.0.7   # Structured logging

# Optional: Additional security (uncomment if needed)
# bleach>=6.1.0             # HTML sanitization (if you allow HTML input)
# flask-wtf>=1.2.0          # CSRF protection
# flask-session>=0.5.0      # Secure session management
# redis>=5.0.0              # For distributed rate limiting
